requirements.txt
tensorflow-gpu==1.12
numpy
opencv-python
scipy
scikit-image
sklearn
pillow